Refactor handlers and add Swagger support
Refactored request handlers to use the new interface `IInvocationHandler` and updated namespaces for better organization. Added Swagger integration for API documentation, including route configuration and metadata. Adjusted controller routes for consistency and clarity.
