+++ sample/RouteDispatcher.API/RequestHandlers/GetCustomMessageHandler.NewApi.cs
@@ -1,11 +1,9 @@
 using RouteDispatcher.API.Requests;
 using RouteDispatcher.Contracts;
 
-namespace RouteDispatcher.API.Handlers
+namespace RouteDispatcher.API.RequestHandlers
 {
-#pragma warning disable CS0618 // Type or member is obsolete
-    public class GetCustomMessageHandler : IRequestHandler<GetCustomMessageRequest, string>
-#pragma warning restore CS0618 // Type or member is obsolete
+    public class GetCustomMessageHandler : IInvocationHandler<GetCustomMessageRequest, string>
     {
         public Task<string> Handle(GetCustomMessageRequest request, CancellationToken cancellationToken)
         {
