+++ sample/RouteDispatcher.API/RequestHandlers/GetCustomMessageHandler.cs
@@ -1,0 +1,14 @@
+// using RouteDispatcher.API.Requests;
+// using RouteDispatcher.Contracts;
+//
+// Old API
+// namespace RouteDispatcher.API.RequestHandlers
+// {
+//     public class GetCustomMessageHandler : IRequestHandler<GetCustomMessageRequest, string>
+//     {
+//         public Task<string> Handle(GetCustomMessageRequest request, CancellationToken cancellationToken)
+//         {
+//             return Task.FromResult($"Custom message: {request.Content}");
+//         }
+//     }
+// }
