+++ sample/RouteDispatcher.API/RouteDispatcher.API.csproj
@@ -8,6 +8,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\..\src\RouteDispatcher\RouteDispatcher.csproj" />
   </ItemGroup>
 </Project>
